Run PostgreSQL command ANALYZE after first insert to table.
Himan relies heavinly on as_grid_v column record_count, which is
populated by the database. In some cases the problem is that database
has not analyzed a table before Himan starts, and the record count
is zero. This causes Himan to pick wrong latest time from database.

This commit modifies grid_to_neons so that if record count of a
table is zero at startup, Himan will run manual ANALYZE after all
fields are inserted.