STU-7007: Fix as_grid column record_count to work with multiple geometries.
From now on record_count column holds only on of two values:
* 0 --> no data
* 1 --> has data

Once this change is in production, and as_grid is populated with correct
information, the view as_grid_v and all the libraries that use record_count
information need to be updated. At that point the manual ANALYZE phase can
also be removed.

This update does not change or brake anything in production, as the column
in the table is not used (the column in the view is used).

Later the data type and name of record_count could be changed.