STU-11448: Add foreign keys to data tables after creation with LIKE
Foreign keys are not transferred with CREATE TABLE .. (LIKE ..).
Fix this by manually fetching fkey information from our template
table and creating those keys based on that information. Postgres 10
does not support foreign keys for partitioned tables (IIRC), but an
update to v12 should be considered.

This commit also changes the default user account for table maintenance
from wetodb to radon_admin. It makes more sense to separate the DDL
commands to another user account which is not used in the same daily
DML commands.