STU-12567: use eccodes to read gribs from memory
Using fmemopen() we can create a FILE pointer to memory
which can then be passed on to eccodes. The upside of this is
that the grib parsing is done by eccodes. The downside is
that the whole file has to fit into memory, and also the
processing is only done after all of file is read resulting
to somewhat slower loading time.